Move global variables to separate file
